# contracts/openapi/orszem-v1.yaml
openapi: 3.1.0
info:
  title: Őrszem Demo v1 API
  version: 1.0.0
  description: >-
    Kanonikus REST/JSON szerződés az Őrszem Demo v1 Public Android és
    Őrszem Szolgálat Android kliensekhez. A public report API anonim;
    minden /service végpont a login kivételével Bearer JWT tokent igényel.
servers:
  - url: /api/v1
    description: API v1 base path
...
paths:
...
  /service/reports/{reportId}:
    get:
      tags: [Service Reports]
      operationId: getServiceReport
      summary: Eset részletei
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Eset részletei
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceReportDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ReportNotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [Service Reports]
      operationId: deleteServiceReport
      summary: Bejelentés végleges törlése (Demo v1.1, csak demo környezet)
      description: >-
        Pilot/teszt adattisztításra szolgáló, kizárólag a demo konfigurációban
        (orszem.demo.deletion-enabled=true, local/demo profil) létező művelet.
        Hitelesített szolgálati kontextust és REPORT_DELETE képességet igényel;
        anonim hívó soha nem törölhet. Végleges (hard) törlés: NEW, IN_PROGRESS
        és ARCHIVED állapotból egyaránt engedélyezett, és az analitika a
        megmaradó sorokat tükrözi. Ha a művelet a környezetben nincs
        engedélyezve, a DELETE handler nem jön létre, így az útvonal 405
        METHOD_NOT_ALLOWED választ ad.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '204':
          description: A bejelentés véglegesen törölve
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/ReportNotFound'
        '405':
          description: >-
            A törlés ebben a környezetben nincs engedélyezve
            (orszem.demo.deletion-enabled=false vagy nem local/demo profil).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                deletionDisabled:
                  value:
                    type: about:blank
                    title: Method not allowed
                    status: 405
                    code: METHOD_NOT_ALLOWED
                    detail: A kért művelet ezen az útvonalon nem engedélyezett.
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ReportId:
      name: reportId
      in: path
      required: true
      description: Report UUID
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: Hiányzó, hibás vagy lejárt access token
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            unauthorized:
              value:
                type: about:blank
                title: Unauthorized
                status: 401
                code: UNAUTHORIZED
                detail: Érvényes szolgálati bejelentkezés szükséges.
    ReportNotFound:
      description: A report nem található
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            notFound:
              value:
                type: about:blank
                title: Report not found
                status: 404
                code: REPORT_NOT_FOUND
                detail: A kért eset nem található.
...
    InternalError:
      description: Nem várt szerverhiba
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    ReportStatus:
      type: string
      enum: [NEW, IN_PROGRESS, ARCHIVED]
...
    EventTypeEmbedded:
      type: object
      additionalProperties: false
      required: [code, categoryCode, categoryLabel, label]
      properties:
        code:
          type: string
          maxLength: 64
        categoryCode:
          type: string
          maxLength: 64
        categoryLabel:
          type: string
          maxLength: 120
        label:
          type: string
          maxLength: 120
    ServiceActorSummary:
      type: object
      additionalProperties: false
      required: [id, displayName]
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
          maxLength: 150
...
    ServiceReportDetail:
      type: object
      additionalProperties: false
      required:
        - id
        - eventType
        - trainIdentifier
        - settlement
        - occurredAt
        - receivedAt
        - status
        - acceptedAt
        - archivedAt
        - acceptedBy
        - archivedBy
      properties:
        id:
          type: string
          format: uuid
        eventType:
          $ref: '#/components/schemas/EventTypeEmbedded'
        trainIdentifier:
          type: string
          maxLength: 64
        settlement:
          type: string
          maxLength: 128
        occurredAt:
          type: string
          format: date-time
        receivedAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/ReportStatus'
        acceptedAt:
          type: [string, 'null']
          format: date-time
        archivedAt:
          type: [string, 'null']
          format: date-time
        acceptedBy:
          oneOf:
            - $ref: '#/components/schemas/ServiceActorSummary'
            - type: 'null'
        archivedBy:
          oneOf:
            - $ref: '#/components/schemas/ServiceActorSummary'
            - type: 'null'
...
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
    ProblemDetails:
      type: object
      additionalProperties: true
      required: [type, title, status, code]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          minLength: 1
          description: Stabil, gépileg feldolgozható Őrszem hibakód
        detail:
          type: string
        instance:
          type: string
        correlationId:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'